refactor: rename QrCodeModal relevant path
